Extract content card component
